separate emb and length

+++ RakutenData.py
@@ -110,8 +110,8 @@ class RakutenData(data.Dataset):
                 ingrs.append(0)
                 validity = False
         ingr_ln = len(ingrs)
-        ingr_ln_tensor = torch.FloatTensor(np.zeros(20))
-        ingr_ln_tensor[ingr_ln] = 1.0
+        ingr_ln_tensor = torch.LongTensor(np.zeros(20))
+        ingr_ln_tensor[ingr_ln] = 1
 
         if self.mode == "use":
             if len(ingrs) < 50:
